refactor to split commands into ai and admin
