+++ src/bot/ai.rs
@@ -1,5 +1,3 @@
-use std::collections::VecDeque;
-
 use crate::{gpt, types, TelitairoBot};
 use teloxide::{prelude::*, utils::command::BotCommands};
 
@@ -8,10 +6,7 @@ use teloxide::{prelude::*, utils::command::BotCommands};
     rename_rule = "lowercase",
     description = "These commands are supported"
 )]
-pub enum Command {
-    #[command(description = "Display this text.")]
-    Help,
-
+pub enum AiCommand {
     #[command(description = "Ask the bot a question.")]
     Ask(String),
 
@@ -19,19 +14,15 @@ pub enum Command {
     Mediate,
 }
 
-pub async fn handle_commands(
+pub async fn handle_ai_commands(
     bot: Bot,
     buffer_store: types::BufferStore,
     telitairo_bot: TelitairoBot,
     msg: Message,
-    cmd: Command,
+    cmd: AiCommand,
 ) -> ResponseResult<()> {
     match cmd {
-        Command::Help => {
-            bot.send_message(msg.chat.id, Command::descriptions().to_string())
-                .await?;
-        }
-        Command::Ask(question) => {
+        AiCommand::Ask(question) => {
             let answer = match gpt::ask(question, telitairo_bot).await {
                 Ok(response) => response,
                 Err(err) => format!("Error getting an answer from OpenAI: {err}"),
@@ -39,7 +30,7 @@ pub async fn handle_commands(
 
             bot.send_message(msg.chat.id, answer).await?;
         }
-        Command::Mediate => {
+        AiCommand::Mediate => {
             let answer = match gpt::mediate(buffer_store, telitairo_bot, msg.chat.id).await {
                 Ok(response) => response,
                 Err(err) => format!("Error getting an answer from OpenAI: {err}"),
@@ -48,29 +39,6 @@ pub async fn handle_commands(
             bot.send_message(msg.chat.id, answer).await?;
         }
     };
-
-    Ok(())
-}
-
-pub async fn handle_messages(
-    buffer_store: types::BufferStore,
-    telitairo_bot: TelitairoBot,
-    msg: Message,
-) -> ResponseResult<()> {
-    let mut buffer_store_lock = buffer_store.write().await;
-    match buffer_store_lock.get_mut(&msg.chat.id) {
-        Some(buffer) => {
-            if buffer.len() == telitairo_bot.buffer_size {
-                buffer.pop_front();
-            }
-            buffer.push_back(msg.clone());
-        }
-        None => {
-            let mut buffer = VecDeque::new();
-            buffer.push_back(msg.clone());
-            buffer_store_lock.insert(msg.chat.id, buffer);
-        }
-    }
 
     Ok(())
 }
